requirements.txt
aiofiles==23.2.1
fastapi==0.115.12
httpx==0.25.2
openai==1.82.1
pydantic-settings==2.9.1
pytest-asyncio==0.21.1
pytest==7.4.3
python-docx==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
uvicorn[standard]==0.24.0